Move class to it own file
